% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\docType{data}
\name{callback}
\alias{callback}
\alias{HipChunkCallback}
\alias{HipSideEffectChunkCallback}
\alias{HipListCallback}
\alias{HipDataFrameCallback}
\title{Callback classes}
\description{
These classes are used to define callback behaviors, and are based
on readr's \code{\link[readr:callback]{readr::callback}} functions.
}
\details{
\itemize{
\item The callbacks \code{HipChunkCallback}, \code{HipListCallback} and
\code{HipSideEffectChunkCallback} should be identical to their readr
counterparts, but have been copied into hipread to ensure that they
work even if readr changes.
\item The callback \code{HipDataFrameCallback} is similar to
readr::DataFrameCallback() except that it uses \code{dplyr::bind_rows()}
instead of \code{rbind()} so that it is faster.
}
}
\keyword{datasets}
\keyword{internal}
