% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipread.R
\docType{package}
\name{hipread-package}
\alias{hipread}
\alias{hipread-package}
\title{hipread: Read Hierarchical Fixed Width Files}
\description{
Read hierarchical fixed width files like those commonly used by 
    many census data providers. Also allows for reading of data in chunks,
    and reading 'gzipped' files without storing the full file in memory.
}
\author{
\strong{Maintainer}: Derek Burk \email{ipums+cran@umn.edu}

Authors:
\itemize{
  \item Greg Freedman Ellis
}

Other contributors:
\itemize{
  \item Joe Grover [contributor]
  \item Mark Padgham [contributor]
  \item Hadley Wickham \email{hadley@rstudio.com} (Code adapted from readr) [contributor]
  \item Jim Hester \email{james.hester@rstudio.com} (Code adapted from readr) [contributor]
  \item Romain Francois (Code adapted from readr) [contributor]
  \item R Core Team (Code adapted from readr) [contributor]
  \item RStudio (Code adapted from readr) [copyright holder, funder]
  \item Jukka Jylänki (Code adapted from readr) [contributor, copyright holder]
  \item Mikkel Jørgensen (Code adapted from readr) [contributor, copyright holder]
  \item University of Minnesota [copyright holder]
}

}
\keyword{internal}
