% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hipread_example}
\alias{hipread_example}
\title{Get path to hipread's example datasets}
\usage{
hipread_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\value{
The filepath to an example file, or if path is empty, a vector of all
available files.
}
\description{
Get access to example extracts.
}
\examples{
hipread_example() # Lists all available examples
hipread_example("test-basic.dat") # Gives filepath for a basic example
}
