% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readh_freqs.R
\name{readh_freqs}
\alias{readh_freqs}
\title{Calculate frequencies from fixed width file without loading into memory}
\usage{
readh_freqs(file, var_names, rt_start, rt_width, var_pos_info,
  compression = NULL, progress = show_progress())
}
\arguments{
\item{file}{A filename}

\item{var_names}{A character vector of names for the variables}

\item{rt_start}{An integer indicating the start position of the rectype
variable (1 indicates the first character)}

\item{rt_width}{An integer indicating the width of the rectype variable,
if there's only one record type, this width can be 0.}

\item{var_pos_info}{A list of fixed width file definitions by rectype,
if there's only one rectype, no name is required, otherwise the name
indicates the value for the rectype variable that indicates the row
is this rectype. Each of these lists must contain a list with equal
length vectors of \code{start} to indicate the starting positions starting
at 1, \code{width} to indicate the widths, and \code{var_pos} which indicates
what order in the data.frame the variable is located (again starting
at 1).}

\item{compression}{If \code{NULL}, guesses the compression from the
file extension (if extension is "gz" uses gzip, otherwise
treats as plain text), can specify it with a string ("txt"
indicates plain text and "gz" for gzip).}

\item{progress}{A logical indicating whether progress should be
displayed on the screen, defaults to showing progress unless
the current context is non-interactive or in a knitr document or
if the user has turned off readr's progress by default using
the option \code{options("readr.show_progress")}.}
}
\value{
A list of frequencies
}
\description{
Calculate the frequency of values in all variables in a fixed width file.
Does so without holding the whole data in memory or creating a full
R data.frame and calling R code on interim pieces. (Probably only
useful inside IPUMS HQ).
}
\keyword{internal}
