% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{col_character}
\alias{reexports}
\alias{col_double}
\alias{reexports}
\alias{col_integer}
\alias{reexports}
\alias{cols}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{readr}{\code{\link[readr]{col_character}}, \code{\link[readr]{col_double}}, \code{\link[readr]{col_integer}}, \code{\link[readr]{cols}}}
}}

