% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readh.R
\name{readh_long}
\alias{readh_long}
\title{Read a hierarchical data file to long format}
\usage{
readh_long(file, var_names, var_types, rt_start, rt_width, var_pos_info,
  var_opts, compression = NULL, progress = show_progress())
}
\arguments{
\item{file}{A filename}

\item{var_names}{A character vector of names for the variables}

\item{var_types}{A character vector of types for the variables. Allowed
types are \code{"character"}, \code{"double"} and \code{"integer"}.}

\item{rt_start}{An integer indicating the start position of the rectype
variable (1 indicates the first character)}

\item{rt_width}{An integer indicating the width of the rectype variable}

\item{var_pos_info}{A list of fixed width file definitions by rectype,
each object in the list must be equal to a value in the rectype variable,
and each of these lists must contain a list with equal lenght vectors of
\code{start} to indicate the starting positions starting at 1, \code{width} to
indicate the widths, and \code{var_pos} which indicates what order in the
data.frame the variable is located (again starting at 1).}

\item{var_opts}{A vector of lists with variable specific options. For
string variables, a logical indicator named \code{trim_ws} is allowed
(if missing it defaults to \code{TRUE}) and for double variables an
integer named \code{imp_dec} indicating the number of implicit decimals
in the variable (defaults to 0).}

\item{compression}{If \code{NULL}, guesses the compression from the
file extension (if extension is "gz" uses gzip, otherwise
treats as plain text), can specify it with a string ("txt"
indicates plain text and "gz" for gzip).}

\item{progress}{A logical indicating whether progress should be
displayed on the screen, defaults to showing progress unless
the current context is non-interactive or in a knitr document or
if the user has turned off readr's progress by default using
the option \code{options("readr.show_progress")}.}
}
\value{
A \code{tbl_df} data frame`
}
\description{
Analagous to \code{readr::read_fwf()} but allowing for
hierarchical fixed width data files (where the data file has rows of
different record types, each with their own variables and column
specifications). Reads the data into "long" format, meaning that
there is one row per observation, and variables that don't apply to
the current observation receive missing values.
}
\examples{

}
